Bugfixing in displaysystem

+++ header/BookSearch.hpp
@@ -8,9 +8,9 @@ class BookSearch {
  private:
 
  public:
-	virtual std::vector<Content*> SearchBooks(std::unordered_map<int, Content*>, std::string);
+	virtual std::vector<Content*> SearchBooks(std::unordered_map<long long, Content*>, std::string);
 	virtual std::vector<Content*> SearchBooks(std::vector<CheckOutData*> std::string);
-	std::vector<Content*> MapToVector(std::unordered_map<int, Content*> catalogue)
+	std::vector<Content*> MapToVector(std::unordered_map<long long, Content*> catalogue)
 	{
 		std::vector<Content*> mapped;
 		for(auto i : catalogue)
