+++ header/BookSort.hpp
@@ -2,15 +2,15 @@
 #define __BOOKSORT_H
 #include <vector>
 #include <string>
-#include "../src/Book.cpp"
+#include "../header/Book.hpp"
 
 class BookSort {
  private:
 
  public:
-	virtual std::vector<Content*> SortBooks(std::unordered_map<int, Content*>) = 0;
+	virtual std::vector<Content*> SortBooks(std::unordered_map<long long, Content*>) = 0;
 	virtual std::vector<Content*> SortBooks(std::vector<CheckOutData*>) = 0;
-	std::vector<Content*> MapToVector(std::unordered_map<int, Content*> catalogue)
+	std::vector<Content*> MapToVector(std::unordered_map<long long, Content*> catalogue)
 	{
 		std::vector<Content*> mapped;
 		for(auto i : catalogue)
